pyproject.toml
[project]
name = "medicine_ai_agent"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-text-splitters>=0.3.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pymilvus>=2.6.7",
    "redis>=5.0.0",
    "rq>=1.16.0",
    "langchain-milvus>=0.3.3",
    "pytest>=9.0.2",
    "pypdf>=6.6.2",
    "python-docx>=1.1.2",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.5",
    "xlrd>=2.0.1",
    "Pillow>=10.3.0",
    "filetype>=1.2.0",
    "python-dotenv>=1.0.1",
    "unstructured>=0.18.31",
    "Markdown>=3.10.1",
    "tiktoken>=0.12.0",
    "httpx>=0.28.1",
    "loguru>=0.7.2",
    "ipython>=9.10.0",
    "langsmith>=0.6.4",
    "pandas>=3.0.0",
    "pandas-stubs>=3.0.0.260204",
    "pymongo>=4.16.0",
    "PyYAML>=6.0.3",
    "websockets>=14.0",
]
